Work on issue #10 and enhancing MessageItem
I was mostly focused on fixing method names, but got off track when I was looking at the MessageItem class.

I decided to make it more robust and readable. I added documentation via docstrings and offloaded repetitive operations to a couple of helper methods. I did away with the status arguments preferring boolean values instead. The reason is because it shifts possible errors from logic errors that would only appear during runtime (and likely only when doing integration tests to syntax errors which are much easier to track down. I added type hints and return types where I could.

The way Most of the set_x_response methods now work is that they either have a was_successful boolean argument that is required, or they take optional argument (like account info data). In that case if the optional argument is `None` or otherwise can't provide the needed information for a successful response, it will set to a failure response with a default failure message if one is not provided.

I tried to make these changes so that we wouldn't have to change too much, but it could be refined further.