Lots of cleanup + global_logger update
I focused on fixing typos, renaming, addressing warnings, adding context managers to keep implementation code easier to follow, and updating the global_logger module to be a bit more useful and cleaner & easier to use.

I made the logger create the directiries and log file if it dosen't already exist and I replaced a bunch of the lambda logging stuff with decorators that log a dictionary for the call that contains not only the function name, but its arguments, return value, any error/exception it raised, it's self-object, and its run-time. I added an optional argument named `level` that should allow you to set the level that the call log will be writen with, but it will default to `DEBUG`.

There are actually 3 decorators currently:
- `@logged_method`
- `@logged_function`
- `@logged_class_method`

Note that the `@logged_class_method` decorator will convert the method to a `classmethod`, but IDEs think it is a mistake. I've tested it using a `unittest` `TestCase` calling it as a class method and it does work as expected. Note that my unittest suite will append to the normal log and assumes no other threads will access the log while the tests are running. I thought about setting up a mock log for testing but didn't invest the time at the moment. For now, only run the unittest if the server is not running and it will be be ok.

Part of the context manager thing is me wanting to experiment with it, but I do think that it significantly simplifies the implementation code where it is utilized and hides a lot of the repetitive complexity within the context manager. To see what I'm talking about please look at db.py. I created a private class that represents the context of the DB transaction, a context manager that sets up the connection and closes it down, and then you just use a with block to do the work utilizing the context object's fields and methods as necessary. I'm really inexperienced with context managers, but I'm really starting to like what you can do with them. I'm sure there are probably ways to improve the design to increase readability and make its use more intuitive.

On the topic of cleanup, I removed a bunch of unused imports while doing a massive cleanup task. If you plan to use an import, I recommend either importing it only when it is actually being utilized (preferable) or paste `# noinspection PyUnresolvedReferences` directly above the unused import to indicate that you don't want it removed during cleanup even though it is unused.

Because of how my code inspection stuff works on my IDE, I can't disable unused import inspection warnings without also disabling a critical set of unresolved reference warnings as well. Those unresolved reference warnings are the most important inspection tool that I use when doing large scale name conversions to make sure I don't have any bad references slip through. It would be really easy to miss one or two of those when my code inspection results have it hidden in a collection of a ton of unused import warnings. It's safer for me to address every warning than risk missing one of these critical ones because it will cause a runtime error if not fixed. I would run our test suite rather than to the exhaustive work of fixing all the warnings, but I don't have access to the servers anymore, so this is the best I can do.

If I removed an import you where planning to use, please just add it back in when you need to use it. I apologize for any inconvenience.