Undo something I tried.
It doesn't do much to load project path from a json stored within the project's directory because you need to be able to find it to use it, which means you already found the project directory.

I was having issues because the service was starting with a working directory of the file system root. I since figured out why I had the issue and corrected it.