added stuff for IDE detection
I wanted to embed the call to warn within the decorator but my IDE only detects it if warn is called from the function definition.

Additionally, if you try to use a variable, function call, or fstring in the arguments for warn, stuff doesn't work as expected.

If we find a cleaner way to get IDE recognition, I'll be happy to change things over at that point, but for the moment I want to get the deprecation warning in my IDE even if it is not the cleanest.